refactor : selectJdbc 와 jdbc 합침
